Stashing work in progress to fully seperate all helper_simple() crutches in victim process. helper_simple() now works with only a dependency on the flush_condition crutch. For addresses far away (and likely soon to be evicted from the TLB), a gadget is needed to cause trigger a minor page fault, pulling the page table into the TLB. Minor page faults inevitabely take longer than a simple memory miss; thus Intel predicts a 0 for the faulting load, allowing speculative execution to continue. This corrupts the expected of 'speculate()' with false loads and must be actively combated.
Two types of gadgets are emulated:
1) Emulation of a gadget which touches the page next to a secret.
2) Emulation of a gadget which touches a page for an arbitrary VA.
In either case, any use of this gadget is always printed to the victim's console for transparency.
It is not unreasonable to assume the attacker could trigger an operation which safely operates on (or near) useful data, but does not provide a means to leak any of this data.
Both of these touch_* functions only load a single byte from the begining of the relevant page (allocated or not).
This load ultimately causes a cache line to be loaded, but its 'utility' is restricted by only ever loading the first cache line in any given page.